+++ src/components/datepicker/__docs__/datepicker.stories.tsx
@@ -14,6 +14,3 @@ const Template: ComponentStory<typeof DatePicker> = args => <DatePicker
   maxDate={args.maxDate && new Date(args.maxDate)}
 />;
 export const Basic = Template.bind({});
-Basic.args = {
-  type: 'dateAndTime'
-}