+++ src/components/popover/reposition.tsx
@@ -1,6 +1,7 @@
 import {Placement} from '@popperjs/core';
+import maxSize from 'popper-max-size-modifier';
 import React, {FC, useState} from 'react';
-import {usePopper} from 'react-popper';
+import {Modifier, usePopper} from 'react-popper';
 
 import {usePopoverContext} from './popoverContext';
 
@@ -18,6 +19,26 @@ export interface RepositionProps {
  * Component.
  */
 
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const applyMaxSize: Modifier<any> = {
+  name: 'applyMaxSize',
+  enabled: true,
+  phase: 'beforeWrite',
+  requires: ['maxSize'],
+  fn({state}) {
+    // The `maxSize` modifier provides this data
+    const {width, height} = state.modifiersData.maxSize;
+
+    // eslint-disable-next-line no-param-reassign
+    state.styles.popper = {
+      ...state.styles.popper,
+      maxWidth: `${width}px`,
+      maxHeight: `${height}px`,
+      width: '100%'
+    };
+  }
+};
+
 export const Reposition: FC<RepositionProps> = props => {
   const {placement, children} = props;
   const popoverContext = usePopoverContext();
@@ -26,6 +47,19 @@ export const Reposition: FC<RepositionProps> = props => {
   const {styles, attributes} = usePopper(popoverContext?.anchor, popperElement, {
     placement,
     modifiers: [
+      {
+        ...maxSize,
+        options: {
+          padding: {
+            top: 10,
+            left: 10,
+            bottom: 10,
+            right: 10
+          },
+          boundary: window.document.body
+        }
+      },
+      applyMaxSize,
       {
         name: 'preventOverflow',
         options: {